v1: Handle recreating engines in the same process
It's possible that code doing offline inference may create an `LLM`
object, delete it, and create a new one within the same process. In
that case we need to be very explicit about ensuring the right cleanup
tasks occur rather than assuming garbage collection will result in the
right things at the right time.

The primary fix is that when `del` is called on an `LLM` object, we
explicitly call `shutdown()` on the engine if it has a shutdown
method. This ensures the v1 shutdown code runs.  I also needed to
propogate the shutdown through the v1 LLMEngine.

The v1 engine client shutdown handling needed some changes, as well:

- No longer rely on `atexit` exclusively. Unregister our atexit
  callback if shutdown occurs prior to the process exiting.

- Harden the shutdown code for when it runs in late stages of process
  shutdown. I observed cases when the final garbage collection in the
  process triggered this code, that key imports were no longer
  accessible (`atexit` and `os`), so don't try those operations if
  they are None.

+++ vllm/v1/engine/core_client.py
@@ -159,10 +159,16 @@ class MPClient(EngineCoreClient):
         atexit.register(self.shutdown)
 
     def shutdown(self):
+        # During final garbage collection in process shutdown, atexit may be
+        # None.
+        if atexit:
+            # in case shutdown gets called via __del__ first
+            atexit.unregister(self.shutdown)
+
         # Shut down the zmq context.
         self.ctx.destroy(linger=0)
 
-        if hasattr(self, "proc_handle"):
+        if hasattr(self, "proc_handle") and self.proc_handle:
             # Shutdown the process if needed.
             if self.proc_handle.proc.is_alive():
                 self.proc_handle.proc.terminate()
@@ -178,8 +184,9 @@ class MPClient(EngineCoreClient):
             ]
             for ipc_socket in ipc_sockets:
                 socket_file = ipc_socket.replace("ipc://", "")
-                if os.path.exists(socket_file):
+                if os and os.path.exists(socket_file):
                     os.remove(socket_file)
+            self.proc_handle = None
 
     def __del__(self):
         self.shutdown()
